1.add referee_data_set 2.add key control(except manual control) 3.change vision control way

+++ User/Components/algorithm/user_lib.h
@@ -99,7 +99,7 @@ extern float32_t jump_error(float32_t err, float32_t err_maxValue);
 #define rad_format(Ang) loop_fp32_constrain((Ang), -PI, PI)
 
 //sigmoid插值
-void sigmoidInterpolation(float32_t startValue, float32_t endValue, uint8_t numPoints, float32_t *output);
+void sigmoidInterpolation(float32_t startValue, float32_t endValue, uint16_t numPoints, float32_t *output);
 /******************************/
 #include <stdint.h>
 #include <stdio.h>
